result_processing/serverless/requirements.txt
# Function dependencies, for example:
# package>=version
auth
google-cloud-storage
requests
